// frontend/src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        ApplicationWithConnectionRequest: {
            ApplicationName: string;
            ApplicationDescription?: string | null;
            Version?: string | null;
            Environment?: string | null;
            Tags?: string | null;
            IsApplicationActive?: boolean;
            ConnectionName: string;
            ConnectionDescription?: string | null;
            DatabaseType: components["schemas"]["DatabaseType"];
// ...
        };
// ...
        DatabaseConnectionRequest: {
            /** Format: uuid */
            ApplicationId?: string | null;
            Name: string;
            Description?: string | null;
            Type: components["schemas"]["DatabaseType"];
            Server: string;
            /** Format: int32 */
            Port?: number | null;
            Database?: string | null;
            Username?: string | null;
            Password?: string | null;
            ConnectionString?: string | null;
            ApiBaseUrl?: string | null;
            ApiKey?: string | null;
            AdditionalSettings?: string | null;
            IsActive?: boolean;
        };
        DatabaseConnectionTestRequest: {
            /** Format: uuid */
            ConnectionId?: string | null;
            ConnectionDetails?: components["schemas"]["DatabaseConnectionRequest"];
        };
        DatabaseConnectionUpdateRequest: {
            /** Format: uuid */
            ApplicationId?: string | null;
            Name: string;
            Description?: string | null;
            Type: components["schemas"]["DatabaseType"];
            Server: string;
            /** Format: int32 */
            Port?: number | null;
            Database?: string | null;
            Username?: string | null;
            Password?: string | null;
            ConnectionString?: string | null;
            ApiBaseUrl?: string | null;
            ApiKey?: string | null;
            AdditionalSettings?: string | null;
            IsActive?: boolean;
            /** Format: uuid */
            Id: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}